refactor ui zip import
